move master-slave yaml to master-slave-common
